cloud 5 - заменил цены на BigDecimal

+++ api/src/main/java/ru/gb/web_market/api/dto/ProductDto.java
@@ -1,5 +1,7 @@
 package ru.gb.web_market.api.dto;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private Long id;
@@ -8,7 +10,7 @@ public class ProductDto {
 
     private String title;
 
-    private Double price;
+    private BigDecimal price;
 
     public Long getId() {
         return id;
@@ -34,18 +36,18 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String article, String title, Double price) {
+    public ProductDto(Long id, String article, String title, BigDecimal price) {
         this.id = id;
         this.article = article;
         this.title = title;
